Less scary `ruff format` message
The ruff formatter has stabilized a good bit since when we added the original message "intended only for experimentation" message. It's time for a less scary message to indicate it is much closer to stable usage than it used to be.

I'm not particular about the specific wording, feel free to change the text to something better.

![image](https://github.com/astral-sh/ruff/assets/6826232/225db573-bfda-446f-a403-a577ee270a0b)