Update configuration.md
The previous configuration for `ruff` contained an unrecognized field `magic-trailing-comma` set to "respect". As of version 0.1.2 of `ruff`, this field was not recognized and resulted in a TOML parse error when running the `ruff format .` command. This change removes the `magic-trailing-comma` field and adds the recognized `skip-magic-trailing-comma` field set to `false`.