ruff_linter: assert that TimeoutErrorAlias only runs on 3.10+
Without this assert, the lint could falsely trigger a safe fix
that is actually unsafe if the higher level logic calling the lint
changes. That is, right now, the lint is only invoked for 3.10+
so everything is fine, but nothing is stopping that from changing.
So we make the assumption clear: if it changes, then we should get
a pretty loud panic.